Call totem subprocess asynchronously
Totem 2.28 seems to run the commandline tool for controlling an
already running Totem instance synchronously, which causes a dead lock
when the main Totem instance itself runs 'totem' (as done by the Remuco
plugin). Now the call has been implemented asynchronously in the Totem
Remuco plugin to prevent this lock. A drawback is that errors during the
call to 'totem' do not get catched but end up in standard out.

Fixes issue 62