Internal canonical player name representation
For now this only affects the Amarok 1.4 adapter. In the internal
canonical name of a player ` ` and `.` characters get removed. The
internal name should be as plain as possible.

Percisely this means, 'amarok1.4' and 'amarok 1.4' is now written as
'amarok14'. On the client the user still sees the name as 'Amarok 1.4'.

--HG--
rename : adapter/amarok1.4/install-check.py => adapter/amarok14/install-check.py
rename : adapter/amarok1.4/remuco-amarok1.4 => adapter/amarok14/remuco-amarok14