Port: lazy dbus handling
This simplifies some of the porting changes before. DBus imports in
mpris.py and manager.py are wrapped by a try-catch block to make them
importable on non-Linux (better: non-DBus) systems. In case DBus is not
available, the code using DBus inside these modules is simply expected
no to get called - which should be true in practice.

This commit is based on changes of Catalin Patulea's fork at
http://github.com/cpatulea/remuco.

--HG--
branch : windows