Add basic 'test shell' functionality
... which is not quite a shell.
This code is more of a proof-of-concept, to allow tests in the adapter without
a client connected.

Update issue 88
It will print a menu with some functions you can call via gobject.idle_add

To use the test shell, just call testshell.setup(playeradapter_instance) before
calling Manager.run().
Then, whenever you want, issue a SIGHUP signal:
$ kill -s 1 <remuco's pid>

It will use the adapter's stdout, so this won't work if you're running in the
background or as a service.