Test shell: cast to int number-ish arguments
The test shell passes the arguments as strings. Since some (most?) of the
parameters in ctrl_* functions are integers, cast everything purely numeric to
int.

After this commit, trivial commands such as ctrl_volume(1) (would be issued as
FUNCTION_NUMBER 1 in the testshell) should work the way they are supposed to.
I haven't tested calls with mixed parameters yet.