Repo: update ignore file
New ignore is 'wiki'. The idea is that the project wiki repository is
cloned to that directory. In doing so the wiki generator scripts can
assume that an up-to-date Remuco source repository is at '..'.

--HG--
extra : transplant_source : %82%0E%DE%CA%C3%08%7Bs%C8%82%DB%A3%B21%F6%14%8A%1D%ED%17